HF-DEV -para el JOBS

+++ main-app/compartido/ajax-definitivas-registrar.php
@@ -18,7 +18,9 @@ $num = mysqli_num_rows($consulta);
 $rB = mysqli_fetch_array($consulta, MYSQLI_BOTH);
 // echo $num; exit();
 if($num==0){
-	mysqli_query($conexion, "DELETE FROM academico_boletin WHERE bol_id='".$rB[0]."'");
+	if(!empty($rB[0])){
+		mysqli_query($conexion, "DELETE FROM academico_boletin WHERE bol_id='".$rB[0]."'");
+	}
 
 	mysqli_query($conexion, "INSERT INTO academico_boletin(bol_carga, bol_estudiante, bol_periodo, bol_nota, bol_tipo, bol_observaciones)VALUES('".$_POST["carga"]."','".$_POST["codEst"]."','".$_POST["per"]."','".$_POST["nota"]."', 4, 'Colocada DEF. por docente.')");
 }else{
